\name{agsc}

\alias{agsc}

\docType{data}

\title{Alabama, Georgia, and South Carolina County Boundaries}

\description{
County polygons for Alabama, Georgia, and South Carolina.
}

\usage{
agsc
}

\format{
A data frame with coordinates of the vertices of the county polgons
\code{(x, y)} and a fabricated variable \code{z}.
}

\examples{
\donttest{data(agsc)}
}

\keyword{datasets}
