\name{BuildHistogram}

\alias{BuildHistogram}

\title{Set Histogram Input Parameters}

\description{
A \acronym{GUI} for specifying input parameters of the \code{\link{hist}}
function.
}

\usage{
BuildHistogram(d, var.names = NULL, var.default = 1L, parent = NULL)
}

\arguments{
\item{d}{something.}
\item{var.names}{character; something.}
\item{var.default}{something.}
\item{parent}{tkwin; the \acronym{GUI} parent window.}
}

\details{
Something
}

\value{
Something
}

\author{J.C. Fisher}

\seealso{
\code{hist}
}

\examples{
\donttest{
plot("something")
}
}

\keyword{misc}
