% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckEntry.R
\name{CheckEntry}
\alias{CheckEntry}
\title{Control Content in Entry Widget}
\usage{
CheckEntry(obj.class, ent.str = "")
}
\arguments{
\item{obj.class}{character.
Name of object class, either \var{real}, \var{integer}, or \var{logical}}

\item{ent.str}{character.
Value from entry widget}
}
\value{
Returns a character string that can be easily converted to the desired object class.
}
\description{
This function enforces content control on entry widgets.
}
\examples{
CheckEntry("numeric", "3.14ab")
CheckEntry("integer", "3.")

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{manip}
