\name{confluence}

\alias{confluence}

\docType{data}

\title{Sample Project}

\description{
An \pkg{RSurvey} project containing list objects \code{\link{srvy.dat}} 
and \code{\link{tran.dat}}.
}

\usage{data(confluence)}

\source{
The survey data was collected within the San Joaquin River Basin, CA, USA, 
at the confluence of the Merced and San Joaquin river in July, 2007.
}

\author{Fisher, J. C.}

\examples{
data(confluence)
summary(srvy.dat())
summary(tran.dat())
}

\keyword{datasets}
