% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.R
\name{Format}
\alias{Format}
\title{Build C-Style String Formats}
\usage{
Format(sample = pi, fmt = "", parent = NULL)
}
\arguments{
\item{sample}{logical, integer, numeric, character, or factor.
Sample value}

\item{fmt}{character.
Conversion specification format, see \code{\link{sprintf}}.}

\item{parent}{tkwin.
\acronym{GUI} parent window}
}
\value{
Returns a character string.
}
\description{
A \acronym{GUI} for the system \code{sprintf} C-library function.
}
\examples{
\dontrun{
  Format(sample = pi, fmt = "\\\%3.8f")
  Format(sample = 3L)
  Format(sample = TRUE)
  Format(sample = "string")
}

}
\seealso{
\code{\link{format}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{misc}
