\name{geoProj}

\alias{geoProj}

\title{Geographical Projection}

\description{
A description of the geographical projection for the current project.
}

\usage{geoProj(tt1 = NULL)}

\arguments{
\item{tt1}{a Tk toplevel widget.}
}

\value{
A character string within the \code{projection} component of \code{\link{srvy.dat}}.
}

\author{Fisher, J. C.}

\examples{
\donttest{geoProj()}
}

\keyword{misc}
