% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportDataset.R
\name{ImportDataset}
\alias{ImportDataset}
\title{Import Data from Package Dataset}
\usage{
ImportDataset(classes = NULL, parent = NULL)
}
\arguments{
\item{classes}{character.
The object classes of data sets that can be loaded.
Set to \code{NULL} to enable loading for all object classes.}

\item{parent}{tkwin.
\acronym{GUI} parent window}
}
\value{
Returns an object of list class with the following components:
  \item{d}{table data}
  \item{src}{vector of length 3 that includes the dataset name, package name, and access date.}
}
\description{
A \acronym{GUI} for importing data from selected \R package datasets.
}
\examples{
\dontrun{
  ans <- ImportDataset(c("data.frame", "matrix"))
}

}
\seealso{
\code{\link{data}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{IO}
