% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportPackage.R
\name{ImportPackage}
\alias{ImportPackage}
\title{Import Dataset from R Package}
\usage{
ImportPackage(classes = NULL, parent = NULL)
}
\arguments{
\item{classes}{character.
The object classes of data sets that can be loaded.
Set to \code{NULL} to enable loading for all object classes.}

\item{parent}{tkwin.
\acronym{GUI} parent window}
}
\value{
Returns an object of list class with the following components:
  \item{d}{table data}
  \item{src}{vector of length 3 that includes the dataset name, package name, and access date.}
}
\description{
A \acronym{GUI} for loading selected dataset types from \R packages.
}
\examples{
\dontrun{
  ans <- ImportPackage(c("data.frame", "matrix"))
}

}
\seealso{
\code{\link{data}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{IO}
