\name{loadPackages}

\alias{loadPackages}

\title{Load Required Packages}

\description{
This function loads \R packages required by \pkg{RSurvey}. 
If a required package is unavailable on the local computer an attempt is made to 
acquire the package from \acronym{CRAN} using an existing network connection.
}

\usage{
loadPackages()
}

\author{Fisher, J. C.}

\examples{
loadPackages()
}

\keyword{misc}
