\name{plotPoints}

\alias{plotPoints}

\title{Plot Spatial Locations and Corresponding State Variable}

\description{
Draw a scatter plot of spatial coordinates (x,y) and their associated state variable (z) values. 
A key showing how the colors map to state variable values is shown to the right of the plot.
}

\usage{
plotPoints()
}

\author{Fisher, J. C.}

\seealso{
\code{\link{plot}}
}

\examples{
data(confluence)
plotPoints()
}

\keyword{hplot}
