\name{plotTime}

\alias{plotTime}

\title{Plot the Raw and Corrected State Variable vs. Time}

\description{
A scatter plot is drawn of raw and corrected state variable measurements versus time.
}

\usage{
plotTime()
}

\details{
This plotting routine is designed as a tool for error correction.
}

\author{Fisher, J. C.}

\seealso{
\code{\link{plot}}
}

\examples{
## None at this time
}

\keyword{hplot}
