\name{project}

\alias{project}

\docType{data}

\title{Example Project Data Set}

\description{
This is an example project data set for \pkg{RSurvey}.
}

\usage{
project
}

\format{
A list containing typical components in a project; these components are
described throughout this documentation and include: data sets,
graphical parameters, and options to control the way in which
\pkg{RSurvey} processes data.
}

\examples{
\donttest{data(project)
# To load project data into RSurvey
Data(replace.all = project)
}
}

\keyword{datasets}
