\name{RSurvey-package}

\alias{RSurvey-package}

\docType{package}

\title{
Analysis of Spatially Distributed Data
}

\description{
This package is a processing program for spatially distributed data.  The application is capable of 
error corrections and data visualization.  A graphical user interface (\acronym{GUI}) is provided.
}

\details{
  \tabular{ll}{
    Package: \tab RSurvey\cr
    Type: \tab Package\cr
    Version: \tab 0.4.3\cr
    Date: \tab 2009-02-28\cr
    License: \tab GPL version 2 or newer.
  }
}

\note{
\R needs to operate as an \acronym{SDI} application with multiple top-level 
windows for the console, graphics and pager. 
Preservation of unit consistency within \pkg{RSurvey} is tasked to the user. 
To assist with unit consistency parameter dimensions are included within the 
\acronym{GUI} entry labels.

}

\author{
Jason C. Fisher

Maintainer: Jason C. Fisher <\email{jfisher@usgs.gov}>
}

\examples{
\donttest{library(RSurvey)}
}

\keyword{package}

