% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search.R
\name{Search}
\alias{Search}
\title{Search}
\usage{
Search(is.replace = FALSE, defaults = NULL, parent = NULL)
}
\arguments{
\item{is.replace}{logical.
If true, the replace component is included.}

\item{defaults}{list.
See \sQuote{Value} section}

\item{parent}{tkwin.
\acronym{GUI} parent window}
}
\value{
Returns an object of list class with the following components:
    \item{find.what}{string to search for}
    \item{replace.with}{replacement string}
    \item{is.match.word}{indicates whether matches be restricted to whole words only.}
    \item{is.match.case}{indicates whether the search is case sensitive.}
    \item{is.reg.exps}{if true, the search is made using \link{regular expression};
      that is, a pattern that describes a set of strings.}
    \item{is.search.col}{indicates whether the search is limited to a single column.}
    \item{is.perl}{indicates whether Perl style regular expressions should be used.}
    \item{is.replace.first}{indicates whether to replace for only the first instance.}
    \item{is.search.sel}{indicates whether the search limited to selected cells.}
}
\description{
A \acronym{GUI} for establishing find and replace arguments in a data table.
}
\examples{
\dontrun{
  ans <- Search()
}

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{misc}
