\name{SetPlotAnnotation}

\alias{SetPlotAnnotation}

\title{Set Plot Annotation}

\description{
A \acronym{GUI} for specifying labels to add to a plot.
}

\usage{SetPlotAnnotation(parent = NULL)}

\arguments{\item{parent}{tkwin; the \acronym{GUI} parent window.}}

\value{
Queries and sets the following components of \code{\link{Data}}:
\item{credit}{character; something.}
\item{explanation}{character; something.}
\item{legend.title}{character; something.}
\item{legend.subtitle}{character; something.}
}

\author{J.C. Fisher}

\examples{\dontrun{SetPlotAnnotation()}}

\keyword{misc}
