\name{srvy.pref}

\alias{srvy.pref}

\title{Set Data Preferences}

\description{
A \acronym{GUI} for specifying the parameters and flags assoicated with error correction, offsets, bathymetry, and interpolation.
}

\usage{srvy.pref(parent)}

\arguments{
\item{parent}{tkwin; the parent window (optional)}
}

\author{Fisher, J. C.}

\examples{
\donttest{srvy.pref()}
}

\keyword{misc}
