\name{tran}

\alias{tran}

\title{Transect Management}

\description{
This function activates the main \acronym{GUI} for managing transect data.
}

\usage{tran(tt1 = NULL)}

\arguments{
\item{tt1}{a Tk toplevel widget.}
}

\author{Fisher, J. C.}

\examples{
\donttest{tran()}
}

\keyword{misc}
