% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildHistogram.R
\name{BuildHistogram}
\alias{BuildHistogram}
\title{GUI: Histogram Input Parameters}
\usage{
BuildHistogram(d, var.names = NULL, var.default = 1L,
  processed.rec = NULL, parent = NULL)
}
\arguments{
\item{d}{list, data.frame, matrix, or numeric.
Vector(s) of values for which the histogram is desired.}

\item{var.names}{character.
Names corresponding to each vector (column) in argument \code{d}.}

\item{var.default}{character or integer.
Vector name or index in argument \code{d}.}

\item{processed.rec}{integer.
Vector of record indexes for processed data.}

\item{parent}{tkwin.
\acronym{GUI} parent window}
}
\value{
\code{NULL}
}
\description{
A graphical user interface (\acronym{GUI}) for specifying input parameters for the \code{\link[graphics]{hist}} function.
}
\examples{
\dontrun{
  BuildHistogram(iris)
}

}
\seealso{
\code{\link[graphics:plothistogram]{plot.histogram}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{misc}
