\name{buildPackage}

\alias{buildPackage}

\title{Package Builder}

\description{
This function is for package developers only.  Package builder has been tested on a 
Windows XP Professional operating system (\acronym{OS}) and within the package source 
directory and not within the installed library.
}

\usage{buildPackage}

\value{
The shell commands necessary for building \pkg{RSurvey} are printed to the \R console.  
The user is tasked with copy/pasting the commands into a Windows command shell.
}

\author{Fisher, J. C.}

\examples{
\dontrun{buildPackage()}
}

\keyword{programming}
