\name{buildPackage}

\alias{buildPackage}

\title{Build R Package in Windows OS}

\description{
This function is intended for package developers and has only been tested on a 
Windows operating system (\acronym{OS}) within the package source directory.
}

\usage{buildPackage()}

\value{
Prints shell commands for building an \R package. 
The user is tasked with copy/pasting the commands into a Windows command shell.
}

\author{Fisher, J. C.}

\examples{
\dontrun{buildPackage()}
}

\keyword{programming}
