% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseColor.R
\name{ChooseColor}
\alias{ChooseColor}
\title{Color Picker}
\usage{
ChooseColor(col, parent = NULL)
}
\arguments{
\item{col}{character.
Initial color, see \sQuote{Value} section}

\item{parent}{tkwin.
\acronym{GUI} parent window}
}
\value{
Returns a selected color in terms of its \code{RGB} components,
a string of the form \code{"#RRGGBB"} where each of the pairs
\code{RR}, \code{GG}, \code{BB} consist of two hexadecimal digits
giving a value in the range \code{00} to \code{FF}.
}
\description{
A \acronym{GUI} for selecting a color.
}
\examples{
\dontrun{
  ChooseColor(col = "#669933")
}

}
\seealso{
\code{\link[grDevices]{col2rgb}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{misc}
