% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoosePch.R
\name{ChoosePch}
\alias{ChoosePch}
\title{GUI: Plotting Symbol Picker}
\usage{
ChoosePch(pch = NA, parent = NULL)
}
\arguments{
\item{pch}{numeric or character.
Initial plotting symbol}

\item{parent}{tkwin.
\acronym{GUI} parent window}
}
\value{
Returns an object of class numeric or integer, specifying the selected plotting symbol.
}
\description{
A graphical user interface (\acronym{GUI}) for selecting a plotting symbol to use.
}
\examples{
\dontrun{
  ChoosePch(pch = "+")
}

}
\seealso{
\code{\link[graphics]{points}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{misc}
