% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineInterpGrid.R
\name{DefineInterpGrid}
\alias{DefineInterpGrid}
\title{Define Interpolation Grid}
\usage{
DefineInterpGrid(grid = NULL, parent = NULL)
}
\arguments{
\item{grid}{list.
Interpolation grid, see \sQuote{Value} section.}

\item{parent}{tkwin.
\acronym{GUI} parent window}
}
\value{
Returns an object of class list with the following components:
  \item{opt}{something}
  \item{res}{numeric components \code{x} and \code{y} giving the grid spacing
    along the x- and y-axis of the interpolated surface, respectively.}
  \item{geo}{something}
}
\description{
A \acronym{GUI} for defining the interpolation grid.
}
\examples{
\dontrun{
  DefineInterpGrid()
}

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{misc}
