\name{ex.project}

\alias{ex.project}

\docType{data}

\title{Example Project Data Set}

\description{
This is an example project data set for \pkg{RSurvey}.
}

\usage{
ex.project
}

\format{
A list containing typical components in a project; these components are
described throughout this documentation and include: data sets,
graphical parameters, and options to control the way in which
\pkg{RSurvey} processes data.
}

\examples{
\donttest{data(ex.project)
# To load a project data set into RSurvey
Data(replace.all = ex.project)
}
}

\keyword{datasets}
