\name{geoproj}

\alias{geoproj}

\title{Set Geographical Projection}

\description{
A \acronym{GUI} for describing the geographical projection for the current project.
}

\usage{geoproj(parent, old = NULL)}

\arguments{
\item{parent}{tkwin; the parent window (optional)}
\item{old}{character; a sting describing the current geographical projection}
}

\value{
Returns a character string describing the geographical projection.
}

\author{Fisher, J. C.}

\examples{
\donttest{geoproj()}
}

\keyword{misc}
