% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportSpreadsheet.R
\name{ImportSpreadsheet}
\alias{ImportSpreadsheet}
\title{Import Data from XML Spreadsheet File}
\usage{
ImportSpreadsheet(parent = NULL)
}
\arguments{
\item{parent}{tkwin.
\acronym{GUI} parent window}
}
\value{
Returns an object of list class with the following components:
  \item{d}{table data}
  \item{src}{vector of length 2 that includes the pathname of the spreadsheet file and access date.}
}
\description{
A \acronym{GUI} for loading selected data sets from an Open XML Spreadsheet file (\file{.xlsx}).
}
\examples{
\dontrun{
  ans <- ImportSpreadsheet()
}

}
\references{
The code in this function was derived with permission from Schaun Wheeler's
  \href{https://gist.github.com/schaunwheeler/5825002}{xlsxToR} function, accessed on 2014-01-01.
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{IO}
