\name{LoadPackages}

\alias{LoadPackages}

\title{Load Required Packages for RSurvey}

\description{
This function loads \R packages required by \pkg{RSurvey}.
If a required package is unavailable on the local computer an attempt
is made to acquire the package from \acronym{CRAN},
\url{http://cran.r-project.org/}, using an existing network connection.
}

\usage{
LoadPackages()
}

\author{Fisher, J.C.}

\examples{
LoadPackages()
}

\keyword{misc}
