% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagePackages.R
\name{ManagePackages}
\alias{ManagePackages}
\title{GUI: Package Manager}
\usage{
ManagePackages()
}
\value{
\code{NULL}
}
\description{
This function installs \R packages suggested by \pkg{RSurvey}.
If a suggested package is unavailable on the local computer,
an attempt is made to acquire the package from \href{https://cran.r-project.org/}{CRAN}
using an existing network connection.
}
\examples{
\dontrun{
  ManagePackages()
}

}
\seealso{
\code{\link[utils]{install.packages}}, \code{\link{requireNamespace}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{misc}
