% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rename.R
\name{Rename}
\alias{Rename}
\title{Rename Values in Character Vector}
\usage{
Rename(names = NULL, cur.name = NULL, win.title = NULL, parent = NULL)
}
\arguments{
\item{names}{character.
Vector of character strings}

\item{cur.name}{character.
Sets the combobox value, name must be included in \code{names}.}

\item{win.title}{character.
String to display as the title of the dialog box.}

\item{parent}{tkwin.
\acronym{GUI} parent window}
}
\value{
Returns a character vector with updated values of \code{names}.
}
\description{
A \acronym{GUI} for renaming values in a vector of character strings.
}
\examples{
\dontrun{
  Rename(names = c("Name1", "Name2", "Name3"), cur.name = "Name2")
}

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{misc}
