\name{RSurvey-package}

\alias{RSurvey-package}

\docType{package}

\title{
Analysis of Spatially Distributed Data
}

\description{
This package is a processing program for spatially distributed data.
It features graphing tools, query building, and polygon clipping.
A graphical user interface (\acronym{GUI}) is provided.
}

\note{
The \pkg{RSurvey} \acronym{GUI} requires \R operate as an \acronym{SDI}
application, using multiple top-level windows for the console, graphics,
and pager.

The set of standards used for coding \pkg{RSurvey} is documented in
\href{http://google-styleguide.googlecode.com/svn/trunk/google-r-style.html}{Google's R Style Guide}.
}

\examples{
\donttest{library(RSurvey)}
}

\keyword{package}
