\name{Search}

\alias{Search}

\title{Search}

\description{
A \acronym{GUI} for establishing find and replace arguments.
}

\usage{
Search(is.replace = FALSE, parent = NULL)
}

\arguments{
\item{is.replace}{logical; if \code{TRUE}, the replace component is included.}
\item{parent}{tkwin; the \acronym{GUI} parent window.}
}

\details{
Something
}

\author{J.C. Fisher}

\examples{\donttest{
ans <- Search()
}}

\keyword{misc}
