% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetPlotAnnotation.R
\name{SetPlotAnnotation}
\alias{SetPlotAnnotation}
\title{GUI: Plot Annotation}
\usage{
SetPlotAnnotation(parent = NULL)
}
\arguments{
\item{parent}{tkwin.
\acronym{GUI} parent window}
}
\value{
Queries and sets the following components of \code{\link{Data}}:
  \item{credit}{mapping credit note}
  \item{explanation}{explanation of gridded-data values.}
  \item{legend.title}{title to be placed at the top of the points legend.}
  \item{legend.subtitle}{subtitle to be placed at the top of the points legend.}
}
\description{
A graphical user interface (\acronym{GUI}) for specifying labels to add to a plot.
}
\examples{
\dontrun{
  SetPlotAnnotation()
}

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{misc}
