\name{srvy.pref}

\alias{srvy.pref}

\title{Set Data Preferences}

\description{
A \acronym{GUI} for specifying error correction, interpolation, and water surface parameters and flags.
}

\usage{srvy.pref(parent)}

\arguments{
\item{parent}{tkwin; the parent window (optional)}
}

\author{Fisher, J. C.}

\examples{
\donttest{srvy.pref()}
}

\keyword{misc}
