\name{srvy}

\alias{srvy}

\title{Main Graphical User Interface}

\description{
This function activates the main \acronym{GUI} for \pkg{RSurvey}.
}

\usage{
srvy()
}

\details{
Most functions within \pkg{RSurvey} are accessible within this \acronym{GUI}. Temporal and 
spatial ranges may also be specificed within the \acronym{GUI}.
}

\author{Fisher, J. C.}

\examples{
\donttest{srvy()}
}

\keyword{misc}
